+++ pyp5js/http/web_app.py
@@ -4,9 +4,9 @@ from flask import Flask, render_template, request, send_from_directory
 from slugify import slugify
 
 from pyp5js import commands
-from pyp5js.config import SKETCHBOOK_DIR
+from pyp5js.config import SKETCHBOOK_DIR, PYODIDE_INTERPRETER
 from pyp5js.exceptions import PythonSketchDoesNotExist, SketchDirAlreadyExistException
-from pyp5js.fs import SketchFiles
+from pyp5js.sketch import Sketch
 
 
 app = Flask(__name__)
@@ -18,8 +18,8 @@ def sketches_list_view():
     sketches = []
     for sketch_dir in (p for p in SKETCHBOOK_DIR.iterdir() if p.is_dir()):
         name = sketch_dir.name
-        sketch_files = SketchFiles(name)
-        if sketch_files.has_all_files:
+        sketch = Sketch(name)
+        if sketch.has_all_files:
             sketches.append({
                 'name': name,
                 'url': f'/sketch/{name}/'
@@ -41,7 +41,9 @@ def add_new_sketch_view():
             context['error'] = "You have to input a sketch name to proceed."
         else:
             try:
-                files = commands.new_sketch(sketch_name)
+                # web client only supports pyodide mode for now
+                # TODO: improve post payload to accept a select
+                files = commands.new_sketch(sketch_name, interpreter=PYODIDE_INTERPRETER)
                 template = 'new_sketch_success.html'
                 context.update({
                     'files': files,
@@ -57,11 +59,11 @@ def add_new_sketch_view():
 @app.route('/sketch/<string:sketch_name>/', defaults={'static_path': ''}, methods=['GET', 'POST'])
 @app.route('/sketch/<string:sketch_name>/<path:static_path>')
 def render_sketch_view(sketch_name, static_path):
-    sketch_files = SketchFiles(sketch_name)
+    sketch = Sketch(sketch_name)
 
     error = ''
     if static_path:
-        return _serve_static(sketch_files.sketch_dir, static_path)
+        return _serve_static(sketch.sketch_dir, static_path)
 
     elif request.method == 'POST':
         py_code = request.form.get('py_code', '')
@@ -73,23 +75,25 @@ def render_sketch_view(sketch_name, static_path):
             error = 'You have to define a draw function.'
         else:
             try:
-                ast.parse(py_code, sketch_files.sketch_py.name)
-                sketch_files.sketch_py.write_text(py_code)
+                ast.parse(py_code, sketch.sketch_py.name)
+                sketch.sketch_py.write_text(py_code)
             except SyntaxError as exc:
                 error = f'SyntaxError: {exc}'
 
     if not error:
         try:
-            commands.transcrypt_sketch(sketch_name)
+            commands.compile_sketch(sketch_name)
         except PythonSketchDoesNotExist:
-            return f"There's no sketch in {sketch_files.sketch_dir.resolve()}", 404
+            return f"There's no sketch in {sketch.sketch_dir.resolve()}", 404
 
     context = {
-        'p5_js_url': sketch_files.urls.p5_js_url,
-        'sketch_js_url': sketch_files.urls.sketch_js_url,
-        'sketch_name': sketch_files.sketch_name,
-        'py_code': sketch_files.sketch_py.read_text(),
+        'p5_js_url': sketch.urls.p5_js_url,
+        'sketch_js_url': sketch.urls.sketch_js_url,
+        'sketch_name': sketch.sketch_name,
+        'py_code': sketch.sketch_py.read_text(),
         'error': error,
+        'js_as_module': sketch.config.is_transcrypt,
+        'live_run': sketch.config.is_pyodide,
     }
     return render_template('view_sketch.html', **context)