+++ pyp5js/sketch.py
@@ -1,41 +1,47 @@
 import os
 import re
-import shutil
-from pathlib import Path
-from cprint import cprint
 from collections import namedtuple
 
 from pyp5js import config
+from pyp5js.config.sketch import SketchConfig
 from pyp5js.exceptions import SketchDirAlreadyExistException, InvalidName
 
 
 SketchUrls = namedtuple('SketchUrls', ['p5_js_url', 'sketch_js_url'])
 
 
-class SketchFiles():
+class Sketch:
+    """
+    This class abstracts the sketch filesystem and configuration.
+    Every path propery return pathlib.Path objects.
+    """
     TARGET_NAME = 'target'
     STATIC_NAME = 'static'
 
-    def __init__(self, sketch_name):
+    def __init__(self, sketch_name, interpreter=config.TRANSCRYPT_INTERPRETER, **cfg):
         self.sketch_name = sketch_name
-        self.from_lib = LibFiles()
+        if self.config_file.exists():
+            self.config = SketchConfig.from_json(self.config_file)
+        else:
+            self.config = SketchConfig(interpreter=interpreter, **cfg)
 
     def validate_name(self):
         does_not_start_with_letter_or_underscore = r'^[^a-zA-Z_]'
         contains_non_alphanumerics_except_underscore = r'[^a-zA-Z0-9_]'
         if re.match(does_not_start_with_letter_or_underscore, self.sketch_name) or \
            re.search(contains_non_alphanumerics_except_underscore, self.sketch_name):
-            raise InvalidName(self.sketch_name)
+            raise InvalidName(self)
 
     def create_sketch_dir(self):
         self.validate_name()
 
         if self.sketch_dir.exists():
-            raise SketchDirAlreadyExistException(self.sketch_dir.resolve())
+            raise SketchDirAlreadyExistException(self)
 
         os.makedirs(self.sketch_dir)
         self.static_dir.mkdir()
         self.target_dir.mkdir()
+        self.config.write(self.config_file)
 
     @property
     def sketch_exists(self):
@@ -66,11 +72,24 @@ class SketchFiles():
 
     @property
     def target_sketch(self):
-        return self.sketch_dir.joinpath("target_sketch.py")
+        # TODO: There is a potential major refactoring here that's to create
+        # base sketch classes and specific implementations. One for a TranscryptSketch
+        # and another one for a PyodideSketch. I think the config
+        # attribute strategy can escalate complexity quickly and it
+        # was a bad idea, but have been working so far...
+        # bonus: opens path to a BrythonSketch ;]
+        if self.config.is_transcrypt:
+            return self.sketch_dir.joinpath("target_sketch.py")
+        else:
+            return self.target_dir.joinpath("target_sketch.js")
 
     @property
     def sketch_py(self):
         return self.sketch_dir.joinpath(f'{self.sketch_name}.py')
+
+    @property
+    def config_file(self):
+        return self.sketch_dir.joinpath('.properties.json')
 
     @property
     def target_dir(self):
@@ -85,45 +104,3 @@ class SketchFiles():
             p5_js_url=f"{self.STATIC_NAME}/p5.js",
             sketch_js_url=f"{self.TARGET_NAME}/target_sketch.js",
         )
-
-
-class LibFiles():
-
-    def __init__(self):
-        self.install = Path(__file__).parent
-
-    @property
-    def templates_dir(self):
-        return self.install.joinpath('templates')
-
-    @property
-    def assets_dir(self):
-        return self.install.joinpath('assets')
-
-    @property
-    def static_dir(self):
-        return self.install.joinpath('static')
-
-    @property
-    def pytop5js(self):
-        return self.install.joinpath('pyp5js.py')
-
-    @property
-    def base_sketch(self):
-        return self.templates_dir.joinpath('base_sketch.py.template')
-
-    @property
-    def target_sketch_template(self):
-        return self.templates_dir.joinpath('target_sketch.py.template')
-
-    @property
-    def index_html(self):
-        return self.templates_dir.joinpath('index.html')
-
-    @property
-    def p5js(self):
-        return self.static_dir.joinpath('p5', 'p5.min.js')
-
-    @property
-    def p5_yml(self):
-        return self.assets_dir.joinpath('p5_reference.yml')